Test setup cleanup
